[Core] Fix PlacementGroup wait hangs with infeasible placement group  (#28899)
Timeout wasn't properly applied when having a sync RPC.

+++ cpp/src/ray/runtime/abstract_ray_runtime.cc
@@ -305,7 +305,7 @@ void AbstractRayRuntime::RemovePlacementGroup(const std::string &group_id) {
 }
 
 bool AbstractRayRuntime::WaitPlacementGroupReady(const std::string &group_id,
-                                                 int timeout_seconds) {
+                                                 int64_t timeout_seconds) {
   return task_submitter_->WaitPlacementGroupReady(group_id, timeout_seconds);
 }
